P2P: Fix p2p_ignore_shared_freq=1 when inviting a peer
p2p_ignore_shared_freq=1 was supposed to allow a MCC-capable device to
ignore a preference for using the same channel on multiple interfaces.
However, it was not used when inviting a peer to re-invoke a persistent
group. This case needs special handling since the peer's channel list is
not available to perform channel reselection and the operating channel
indicated in the Invitation Request frames ends up getting used as the
operating channel if the transmitted of that frames becomes the GO.

Signed-hostap: Jouni Malinen <jouni@qca.qualcomm.com>
Git-commit: d0f61a4beafd7b33ae93eeb70c33a7c330924faf
Git-repo : git://w1.fi/srv/git/hostap.git
Change-Id: I40f1002097be4f7f108eb749280025d306f60492
CRs-fixed: 620921